Support sorting by 'random' (#3940)
* Support sorting by 'random'

This adds a new 'random' sort key, which is stable within a single run
of `task` but will produce a different result for each run. This allows,
for example, selecting a random task with `task all limit:1`. The 
configuration needs to be in the `.taskrc` such as `report.<report-name>.sort = random`.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>